proc: generalize escapeCheck and allocString
Generalizes the function for checking for escaping pointers so that it
can be used to iterate on all pointers of a variable.
Also generalizes the string allocation opcodes to be generally usable
to call functions known when the expression to evaluate is known.
These are both refactoring to prepare to use runtime.debugPinner in
call injection.